+++ src/app/index.route.js
@@ -43,6 +43,12 @@
                 templateUrl: 'app/collections/products/products.html',
                 title: 'Decertified Products',
             })
+            .when('/collections/sed', {
+                controller: 'SedCollectionController',
+                controllerAs: 'vm',
+                templateUrl: 'app/collections/sed/sed.html',
+                title: 'SED Information for 2014 &amp; 2015 Edition Products',
+            })
             .when('/collections/transparencyAttestations', {
                 controller: 'TransparencyAttestationsController',
                 controllerAs: 'vm',